BAU: Update gpg45 structure to match with stubs and work with the new vot claim checks
